Optimised functions, changed inputs, made drive speed and turn speed go to dashboard, correctly referenced classes through static functions for threads, Changed arm to be controlled by Dpad

+++ src/main/cpp/Functions.cpp
@@ -1,12 +1,12 @@
 #include "Functions.h"
 
 // Divide input by 12
-double Functions::InputVoltage(double voltage) {
+float Functions::InputVoltage(float voltage) {
     return(voltage/12);
 }
 
 // Set motor at speed for a specific amount of time (ntested)
-void Functions::SetTimedMotor(SpeedController* motor, double motorSpeed, double motorTime) {
+void Functions::SetTimedMotor(SpeedController* motor, float motorSpeed, float motorTime) {
     motor->Set(motorSpeed);
     Wait(motorTime);
     motor->Set(0.0);
